Validate the RMASCA model LMM

This function performs leave-one-out robustness testing of your RMASCA model. If you didn't specify the number of runs nValRuns when initializing the model (see RMASCA), you can do it by running for example model$nValRuns <- 100 prior to calling validate. Your dataset is divided into nValFold partitions, keeping group proportions, and one of these are left out. nValFold is set the same way as nValRuns.

validate(object, participantColumn = FALSE)

Arguments

object
An RMASCA object

participantColumn
The name of the column containing participant identifier. Needed if not set during initialization of the model.

Value

An RMASCA object

Examples

load("PE.Rdata")
model$nValRuns <- 10
model.val <- validate(model, participantColumn = "ID")

#> Running validation...
#> - Run  1  of  10 

#> Error in sanitizeObject(object): The dataframe must contain a column names 'time'